backend/requirements.txt
fastapi 
uvicorn[standard] 
sqlalchemy[asyncio] 
asyncpg 
alembic 
pydantic-settings 
python-dotenv 
httpx 
beautifulsoup4 
openai
python-multipart
